+++ sdk/servicebus/service-bus/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 7.0.0-preview.5 (Unreleased)
 
+- User agent details can now be added to the outgoing requests by passing the user-agent prefixes to the `ServiceBusClient` and the `ServiceBusManagementClient` through options.
+  Example user-agent string if the prefix `SampleApp` is provided to `ServiceBusManagementClient`:
+  `SampleApp azsdk-js-azureservicebus/7.0.0-preview.5 core-http/1.1.5 Node/v12.16.0 OS/(x64-Windows_NT-10.0.18363)`
+  [PR 10092](https://github.com/Azure/azure-sdk-for-js/pull/10092)
 - Adding `deadLetterErrorDescription` and `deadLetterReason` properties on the received messages. Previously, they were under the `properties` in the message.
   OLD: `message.properties["DeadLetterReason"]` and `message.properties["DeadLetterErrorDescription"]`
   NEW: `message.deadLetterReason` and `message.deadLetterErrorDescription`